+++ CoreWars.Engine.TestProject/TokenizerSmokeUnitTestSamples/TokenizerSmokeUnitTestSamplesStorage.cs
@@ -5,53 +5,53 @@ using System.Reflection;
 namespace CoreWars.Engine.TokenizerSmokeUnitTestSamples {
     internal static class TokenizerSmokeUnitTestSamplesStorage {
 
-        public static (string Name, string Cotent, IEnumerable<(int lineNumber, string line)> Codelines) GetExample000() => GetEmbeddedProgram("Example000.redcode");
+        public static (string Name, string Cotent, IEnumerable<(short lineNumber, string line)> Codelines) GetExample000() => GetEmbeddedProgram("Example000.redcode");
 
-        public static (string Name, string Cotent, IEnumerable<(int lineNumber, string line)> Codelines) GetExample001() => GetEmbeddedProgram("Example001.redcode");
-        public static (string Name, string Cotent, IEnumerable<(int lineNumber, string line)> Codelines) GetExample002() => GetEmbeddedProgram("Example002.redcode");
-        public static (string Name, string Cotent, IEnumerable<(int lineNumber, string line)> Codelines) GetExample003() => GetEmbeddedProgram("Example003.redcode");
-        public static (string Name, string Cotent, IEnumerable<(int lineNumber, string line)> Codelines) GetExample004() => GetEmbeddedProgram("Example004.redcode");
-        public static (string Name, string Cotent, IEnumerable<(int lineNumber, string line)> Codelines) GetExample005() => GetEmbeddedProgram("Example005.redcode");
-        public static (string Name, string Cotent, IEnumerable<(int lineNumber, string line)> Codelines) GetExample006() => GetEmbeddedProgram("Example006.redcode");
-        public static (string Name, string Cotent, IEnumerable<(int lineNumber, string line)> Codelines) GetExample007() => GetEmbeddedProgram("Example007.redcode");
-        public static (string Name, string Cotent, IEnumerable<(int lineNumber, string line)> Codelines) GetExample008() => GetEmbeddedProgram("Example008.redcode");
-        public static (string Name, string Cotent, IEnumerable<(int lineNumber, string line)> Codelines) GetExample009() => GetEmbeddedProgram("Example009.redcode");
-        public static (string Name, string Cotent, IEnumerable<(int lineNumber, string line)> Codelines) GetExample010() => GetEmbeddedProgram("Example010.redcode");
+        public static (string Name, string Cotent, IEnumerable<(short lineNumber, string line)> Codelines) GetExample001() => GetEmbeddedProgram("Example001.redcode");
+        public static (string Name, string Cotent, IEnumerable<(short lineNumber, string line)> Codelines) GetExample002() => GetEmbeddedProgram("Example002.redcode");
+        public static (string Name, string Cotent, IEnumerable<(short lineNumber, string line)> Codelines) GetExample003() => GetEmbeddedProgram("Example003.redcode");
+        public static (string Name, string Cotent, IEnumerable<(short lineNumber, string line)> Codelines) GetExample004() => GetEmbeddedProgram("Example004.redcode");
+        public static (string Name, string Cotent, IEnumerable<(short lineNumber, string line)> Codelines) GetExample005() => GetEmbeddedProgram("Example005.redcode");
+        public static (string Name, string Cotent, IEnumerable<(short lineNumber, string line)> Codelines) GetExample006() => GetEmbeddedProgram("Example006.redcode");
+        public static (string Name, string Cotent, IEnumerable<(short lineNumber, string line)> Codelines) GetExample007() => GetEmbeddedProgram("Example007.redcode");
+        public static (string Name, string Cotent, IEnumerable<(short lineNumber, string line)> Codelines) GetExample008() => GetEmbeddedProgram("Example008.redcode");
+        public static (string Name, string Cotent, IEnumerable<(short lineNumber, string line)> Codelines) GetExample009() => GetEmbeddedProgram("Example009.redcode");
+        public static (string Name, string Cotent, IEnumerable<(short lineNumber, string line)> Codelines) GetExample010() => GetEmbeddedProgram("Example010.redcode");
 
-        public static (string Name, string Cotent, IEnumerable<(int lineNumber, string line)> Codelines) GetExample011() => GetEmbeddedProgram("Example011.redcode");
-        public static (string Name, string Cotent, IEnumerable<(int lineNumber, string line)> Codelines) GetExample012() => GetEmbeddedProgram("Example012.redcode");
-        public static (string Name, string Cotent, IEnumerable<(int lineNumber, string line)> Codelines) GetExample013() => GetEmbeddedProgram("Example013.redcode");
-        public static (string Name, string Cotent, IEnumerable<(int lineNumber, string line)> Codelines) GetExample014() => GetEmbeddedProgram("Example014.redcode");
-        public static (string Name, string Cotent, IEnumerable<(int lineNumber, string line)> Codelines) GetExample015() => GetEmbeddedProgram("Example015.redcode");
-        public static (string Name, string Cotent, IEnumerable<(int lineNumber, string line)> Codelines) GetExample016() => GetEmbeddedProgram("Example016.redcode");
-        public static (string Name, string Cotent, IEnumerable<(int lineNumber, string line)> Codelines) GetExample017() => GetEmbeddedProgram("Example017.redcode");
-        public static (string Name, string Cotent, IEnumerable<(int lineNumber, string line)> Codelines) GetExample018() => GetEmbeddedProgram("Example018.redcode");
-        public static (string Name, string Cotent, IEnumerable<(int lineNumber, string line)> Codelines) GetExample019() => GetEmbeddedProgram("Example019.redcode");
-        public static (string Name, string Cotent, IEnumerable<(int lineNumber, string line)> Codelines) GetExample020() => GetEmbeddedProgram("Example020.redcode");
+        public static (string Name, string Cotent, IEnumerable<(short lineNumber, string line)> Codelines) GetExample011() => GetEmbeddedProgram("Example011.redcode");
+        public static (string Name, string Cotent, IEnumerable<(short lineNumber, string line)> Codelines) GetExample012() => GetEmbeddedProgram("Example012.redcode");
+        public static (string Name, string Cotent, IEnumerable<(short lineNumber, string line)> Codelines) GetExample013() => GetEmbeddedProgram("Example013.redcode");
+        public static (string Name, string Cotent, IEnumerable<(short lineNumber, string line)> Codelines) GetExample014() => GetEmbeddedProgram("Example014.redcode");
+        public static (string Name, string Cotent, IEnumerable<(short lineNumber, string line)> Codelines) GetExample015() => GetEmbeddedProgram("Example015.redcode");
+        public static (string Name, string Cotent, IEnumerable<(short lineNumber, string line)> Codelines) GetExample016() => GetEmbeddedProgram("Example016.redcode");
+        public static (string Name, string Cotent, IEnumerable<(short lineNumber, string line)> Codelines) GetExample017() => GetEmbeddedProgram("Example017.redcode");
+        public static (string Name, string Cotent, IEnumerable<(short lineNumber, string line)> Codelines) GetExample018() => GetEmbeddedProgram("Example018.redcode");
+        public static (string Name, string Cotent, IEnumerable<(short lineNumber, string line)> Codelines) GetExample019() => GetEmbeddedProgram("Example019.redcode");
+        public static (string Name, string Cotent, IEnumerable<(short lineNumber, string line)> Codelines) GetExample020() => GetEmbeddedProgram("Example020.redcode");
 
-        public static (string Name, string Cotent, IEnumerable<(int lineNumber, string line)> Codelines) GetExample021() => GetEmbeddedProgram("Example021.redcode");
-        public static (string Name, string Cotent, IEnumerable<(int lineNumber, string line)> Codelines) GetExample022() => GetEmbeddedProgram("Example022.redcode");
-        public static (string Name, string Cotent, IEnumerable<(int lineNumber, string line)> Codelines) GetExample023() => GetEmbeddedProgram("Example023.redcode");
-        public static (string Name, string Cotent, IEnumerable<(int lineNumber, string line)> Codelines) GetExample024() => GetEmbeddedProgram("Example024.redcode");
-        public static (string Name, string Cotent, IEnumerable<(int lineNumber, string line)> Codelines) GetExample025() => GetEmbeddedProgram("Example025.redcode");
-        public static (string Name, string Cotent, IEnumerable<(int lineNumber, string line)> Codelines) GetExample026() => GetEmbeddedProgram("Example026.redcode");
-        public static (string Name, string Cotent, IEnumerable<(int lineNumber, string line)> Codelines) GetExample027() => GetEmbeddedProgram("Example027.redcode");
-        public static (string Name, string Cotent, IEnumerable<(int lineNumber, string line)> Codelines) GetExample028() => GetEmbeddedProgram("Example028.redcode");
-        public static (string Name, string Cotent, IEnumerable<(int lineNumber, string line)> Codelines) GetExample029() => GetEmbeddedProgram("Example029.redcode");
-        public static (string Name, string Cotent, IEnumerable<(int lineNumber, string line)> Codelines) GetExample030() => GetEmbeddedProgram("Example030.redcode");
+        public static (string Name, string Cotent, IEnumerable<(short lineNumber, string line)> Codelines) GetExample021() => GetEmbeddedProgram("Example021.redcode");
+        public static (string Name, string Cotent, IEnumerable<(short lineNumber, string line)> Codelines) GetExample022() => GetEmbeddedProgram("Example022.redcode");
+        public static (string Name, string Cotent, IEnumerable<(short lineNumber, string line)> Codelines) GetExample023() => GetEmbeddedProgram("Example023.redcode");
+        public static (string Name, string Cotent, IEnumerable<(short lineNumber, string line)> Codelines) GetExample024() => GetEmbeddedProgram("Example024.redcode");
+        public static (string Name, string Cotent, IEnumerable<(short lineNumber, string line)> Codelines) GetExample025() => GetEmbeddedProgram("Example025.redcode");
+        public static (string Name, string Cotent, IEnumerable<(short lineNumber, string line)> Codelines) GetExample026() => GetEmbeddedProgram("Example026.redcode");
+        public static (string Name, string Cotent, IEnumerable<(short lineNumber, string line)> Codelines) GetExample027() => GetEmbeddedProgram("Example027.redcode");
+        public static (string Name, string Cotent, IEnumerable<(short lineNumber, string line)> Codelines) GetExample028() => GetEmbeddedProgram("Example028.redcode");
+        public static (string Name, string Cotent, IEnumerable<(short lineNumber, string line)> Codelines) GetExample029() => GetEmbeddedProgram("Example029.redcode");
+        public static (string Name, string Cotent, IEnumerable<(short lineNumber, string line)> Codelines) GetExample030() => GetEmbeddedProgram("Example030.redcode");
 
-        public static (string Name, string Cotent, IEnumerable<(int lineNumber, string line)> Codelines) GetExample031() => GetEmbeddedProgram("Example031.redcode");
-        public static (string Name, string Cotent, IEnumerable<(int lineNumber, string line)> Codelines) GetExample032() => GetEmbeddedProgram("Example032.redcode");
-        public static (string Name, string Cotent, IEnumerable<(int lineNumber, string line)> Codelines) GetExample033() => GetEmbeddedProgram("Example033.redcode");
-        public static (string Name, string Cotent, IEnumerable<(int lineNumber, string line)> Codelines) GetExample034() => GetEmbeddedProgram("Example034.redcode");
-        public static (string Name, string Cotent, IEnumerable<(int lineNumber, string line)> Codelines) GetExample035() => GetEmbeddedProgram("Example035.redcode");
-        public static (string Name, string Cotent, IEnumerable<(int lineNumber, string line)> Codelines) GetExample036() => GetEmbeddedProgram("Example036.redcode");
-        public static (string Name, string Cotent, IEnumerable<(int lineNumber, string line)> Codelines) GetExample037() => GetEmbeddedProgram("Example037.redcode");
-        public static (string Name, string Cotent, IEnumerable<(int lineNumber, string line)> Codelines) GetExample038() => GetEmbeddedProgram("Example038.redcode");
-        public static (string Name, string Cotent, IEnumerable<(int lineNumber, string line)> Codelines) GetExample039() => GetEmbeddedProgram("Example039.redcode");
-        public static (string Name, string Cotent, IEnumerable<(int lineNumber, string line)> Codelines) GetExample040() => GetEmbeddedProgram("Example040.redcode");
+        public static (string Name, string Cotent, IEnumerable<(short lineNumber, string line)> Codelines) GetExample031() => GetEmbeddedProgram("Example031.redcode");
+        public static (string Name, string Cotent, IEnumerable<(short lineNumber, string line)> Codelines) GetExample032() => GetEmbeddedProgram("Example032.redcode");
+        public static (string Name, string Cotent, IEnumerable<(short lineNumber, string line)> Codelines) GetExample033() => GetEmbeddedProgram("Example033.redcode");
+        public static (string Name, string Cotent, IEnumerable<(short lineNumber, string line)> Codelines) GetExample034() => GetEmbeddedProgram("Example034.redcode");
+        public static (string Name, string Cotent, IEnumerable<(short lineNumber, string line)> Codelines) GetExample035() => GetEmbeddedProgram("Example035.redcode");
+        public static (string Name, string Cotent, IEnumerable<(short lineNumber, string line)> Codelines) GetExample036() => GetEmbeddedProgram("Example036.redcode");
+        public static (string Name, string Cotent, IEnumerable<(short lineNumber, string line)> Codelines) GetExample037() => GetEmbeddedProgram("Example037.redcode");
+        public static (string Name, string Cotent, IEnumerable<(short lineNumber, string line)> Codelines) GetExample038() => GetEmbeddedProgram("Example038.redcode");
+        public static (string Name, string Cotent, IEnumerable<(short lineNumber, string line)> Codelines) GetExample039() => GetEmbeddedProgram("Example039.redcode");
+        public static (string Name, string Cotent, IEnumerable<(short lineNumber, string line)> Codelines) GetExample040() => GetEmbeddedProgram("Example040.redcode");
 
-        private static (string Name, string Cotent, IEnumerable<(int lineNumber, string line)> Codelines) GetEmbeddedProgram(string programName) {
+        private static (string Name, string Cotent, IEnumerable<(short lineNumber, string line)> Codelines) GetEmbeddedProgram(string programName) {
             var assembly = Assembly.GetExecutingAssembly();
             var resourceName = $"CoreWars.Engine.TokenizerSmokeUnitTestSamples.{programName}.txt";
 
@@ -62,14 +62,13 @@ namespace CoreWars.Engine.TokenizerSmokeUnitTestSamples {
             }
         }
 
-        private static IEnumerable<(int lineNumber, string line)> ToLines(this string lines) {
+        private static IEnumerable<(short lineNumber, string line)> ToLines(this string lines) {
             using (StringReader stringReader = new StringReader(lines)) {
                 string line;
-                int lineNumber = 1;
+                short lineNumber = 1;
                 while ((line = stringReader.ReadLine()) != null)
                     yield return (lineNumber: lineNumber++, line: line);
             }
-
         }
     }
 }
